Finalização tela de caixas abertos. Ajustes telas campos obrigatórios.

+++ src/main/java/br/com/augustocosta/acs/integration/dto/dtoCaixa.java
@@ -2,6 +2,8 @@ package br.com.augustocosta.acs.integration.dto;
 
 import lombok.*;
 import org.springframework.format.annotation.DateTimeFormat;
+
+import java.math.BigDecimal;
 import java.time.*;
 
 @Getter
@@ -21,13 +23,13 @@ public class dtoCaixa {
     private Integer responsavelFechamento;
     private String nomeRespFechamento;
     private LocalDateTime dataFechamento;
-    private Double valorAbertura;
-    private Double credito;
-    private Double debito;
-    private Double dinheiro;
-    private Double pix;
-    private Double valorProvisorio;
-    private Double valorFechamento;
+    private BigDecimal valorAbertura;
+    private BigDecimal credito;
+    private BigDecimal debito;
+    private BigDecimal dinheiro;
+    private BigDecimal pix;
+    private BigDecimal valorProvisorio;
+    private BigDecimal valorFechamento;
     private Boolean ativo;
     private LocalDateTime dataCriacao;
     private LocalDateTime dataAlteracao;
